+++ NEWS.md
@@ -25,6 +25,7 @@
 - Added support for VHDL-2008 matching `select?` statements (#705).
 - Added support for the new `'designated_subtype` and `'index`
   attributes in VHDL-2019.
+- Implemented the date/time functions from `std.env` in VHDL-2019.
 
 ## Version 1.9.2 - 2023-05-01
 - Fix elaboration errors with recursive entity instantiation (#668).